+++ openwpm/commands/utils/firefox_profile.py
