+++ openwpm/deploy_browsers/deploy_firefox.py
@@ -8,7 +8,7 @@ from pyvirtualdisplay import Display
 from selenium import webdriver
 from selenium.webdriver.firefox.firefox_profile import FirefoxProfile
 
-from ..Commands.profile_commands import load_profile
+from ..commands.profile_commands import load_profile
 from ..errors import BrowserConfigError
 from ..utilities.platform_utils import get_firefox_binary_path
 from . import configure_firefox
