+++ openwpm/deploy_browsers/screen_resolutions.txt
